switching to docs/updated-examples
